Migrated CrawlCtrl.Extensions.Microsoft.DependencyInjection to this repo.
CrawlCtrl.Extensions.Microsoft.DependencyInjection to this repository.

This has been done because to two project will follow each other 1:1. A new release of CrawlCtrl will result in a new release of the DI extension package. By moving the project into this solution, we can automate the release and save time. WIN!