Migrated CrawlCtrl.Reader to this repo.
Moved CrawlCtrl.Reader from a separate solution/repo to this solution/repo.

This has been done since the core CrawlCtrl project are going to follow eachother 1:1. A release of the CrawlCtrl package results in the new release of CrawlCtrl.Reader with the same version. By moving the project into this solution we can automate the release and save time. WINNER WINNER CHICKEN DINNER!