Use immutable options while deserializing
Options specified in RobotsDeserializerOptions are now mapped to an ImmutableRobotsDeserializationOptions instance as the first thing in RobotsDeserializer, before being passed on to the rest of the code. If no options are provided (null) as instance with default options are created before the mapping step.

Minor cleanup in RobotsDeserializerOptions: The properties IncludeSitemaps (bool) and SitemapsInclusionScope (enum) was been consolidated into the unified enum property SitemapPolicy. SitemapPolicy covers all of the same use cases as the two old properties.

Options are now passed into ILineDeserializer instance through the Deserialize method instead of the constructor. This has been done in preparation for custom line deserializer support. Passing the options in through the method instead of the constructor will give 3rd party developers access to the options.